feat: move to util function
