--- packages/ui/src/components/molecule/user-dialog/shortenName.ts
@@ -1,6 +1,0 @@
-export function getShortenName(fullName: string) {
-  const names = fullName.split(' ')
-  const firstName = names[0]
-  const lastName = names[names.length - 1]
-  return `${firstName} ${lastName.charAt(0)}.`
-}
